requirements.txt
altair==4.0
mysql-connector-python==8.0.33
streamlit==1.11.0
vega_datasets
